fixed bullcrap

+++ core/src/com/retrocombine/breakout/Position.java
@@ -5,10 +5,10 @@ package com.retrocombine.breakout;
  */
 public class Position {
 
-    public int x;
-    public int y;
+    public float x;
+    public float y;
 
-    public Position(int x, int y) {
+    public Position(float x, float y) {
         this.x = x;
         this.y = y;
     }
